Preferring use of hiopVector interface
All instances of hiopVectorPar are being replaced with hiopVector, so we
can more easily swap in a new implementation of hiopVector.

There currently seems to be quite a few issues with use of the
implementation instead of the interface, which I'm working through
solving right now. This currently does not build as we need to dynamic
cast to hiopVectorPar to use the implementation-specific methods in
cosmin's algorithms.

I think a solution may be to move the `*_local` methods into the
interface even though they do not belong there, just to make the changes
to the code more incremental. This many dynamic_casts is going to induce
quite an overhead, and is not portable.